You can plunge geysers with a normal plunger and removes reinforced plungers (#2864)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
You can plunger and activate geysers with a standard plunger. 

Also removes reinforced plungers since they don't do anything anymore.
and adds a normal plunger to the Schmiedeberg in it's place.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The reinforced plunger was effectively made unobtainable with #2617
removing it from the vendor stock, aside from the one ship it's
currently mapped on (Schmiedeberg), with no other available sources.
Geysers are basically inaccessible to the vast majority of players as a
result.

Even then, having the reinforced plunger only obtainable in the medical
vendor gated it behind a small subset of ships, and meant it was almost
never seen anyways.

This will make geysers usable again and the increased accessibility
hopefully means it will actually be interacted with.


<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
tweak: Geysers can now be plunged with a normal plunger.
tweak: Schmiedeberg's reinforced plunger replaced with a normal plunger
del: Reinforced plungers
imagedel: Reinforced plunger sprites
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->