Remaps the Shetland (#2774)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

![faroe](https://github.com/shiptest-ss13/Shiptest/assets/60533805/fcae91a3-a13e-457e-8292-23deba44527c)

![faroe_areas](https://github.com/shiptest-ss13/Shiptest/assets/60533805/ed74e3bd-160f-4c36-9839-2d34bd132f7b)

Brings the Shetland up to modern mapping standards, makes the layout
less garbage, adds some amenities and RP facilities, and rebalances the
ship's mechanical features.

- RnD and ORM removed
- TEG added
- Weapons and armor rebalanced
- Massive aesthetic improvements

## Why It's Good For The Game

Shetland is the oldest ship we have left in the game and despite getting
a couple updates along the way it really shows. It was either update it
or kill it, and I think the Shetland still has good potential and is
worth saving.

## Changelog

:cl:
tweak: Remapped the Shetland
balance: Removed RnD and ORM from Shetland
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->