re:magazines (#2867)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Resprites most magazines in the game. PRed on behalf of Imaginos16 and
spookywastaken with some minor edits

## Why It's Good For The Game

Note: some are from some unimplmented guns and are not apart of this PR.

![Screenshot_12202](https://github.com/shiptest-ss13/Shiptest/assets/58402542/a2971de5-2921-4793-9b4c-3ee7f0ce597d)

## Changelog

:cl: Imaginos16, spookywastaken
add: Resprites a ton of gun magazines
fix: Changes how magainzes with only a empty and loaded state are
/:cl: