Brings back old gun suicide_act behavior (#2784)
## About The Pull Request
Brings back old suicide_act on handle_suicide proc of guns.

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->


https://github.com/shiptest-ss13/Shiptest/assets/81882910/a5363807-0b49-46fc-aae6-cc1253cc1dbf

## Changelog

:cl:
add: Suicide & Execution with a damaging firearm is now more deadly and
gruesome.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->