Adds missing descs to Nanotrasen uniform items (#2951)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

The Nanotrasen resprite was mistakenly merged without any item
descriptions written. This PR fixes that.

## Why It's Good For The Game

descriptive text is good and is also the only way to actually put in
representation for the new Nanotrasen factions.

## Changelog

:cl:
fix: fixed missing object descriptions on nanotrasen uniforms
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->