Match bullets no longer have a highly above-average chance to ricochet into their user's foot (#2944)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Nothing is stopping the only person in ricochet range from being the
person shooting the gun. Which looks really stupid if you are trying to
do something funny and shoot yourself in the foot 6 times.

## Why It's Good For The Game

"bullet that shoots the person who shot it" isn't an ammo type

## Changelog

:cl:
tweak: match rounds will no longer consider the person firing them as a
viable target when ricocheting
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Signed-off-by: Theos <theubernyan@gmail.com>