Removes reality integrity issues from EXOCOM (#2938)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The EXO Request
A quantum superposition around the name of the EXO mining company has
been isolated and destroyed by the shiptest reality force.
From hence on, all shall know only the name **EXOCOM**, rather than
also, sometimes, EXOCON.
**EXOCON is not, and was never real.** 
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For EXOCON (Which is not real)
EXOCON is not real. Please report any sightings of EXOCON personnel or
equipment for immediate termination. EXOCON is not real.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Nothing Has Changed. Please Remain Calm.

:cl:
fix: weird uncertainty around the name of EXOCOM. EXOCON is not real. 
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->