Add default return for robotic emotes (#2955)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

PR fixes this bug


![image](https://github.com/shiptest-ss13/Shiptest/assets/32931701/1e7a4f99-e2d6-49a3-bdbc-2953d90a892e)


## Why It's Good For The Game

Bugfixes are good

## Changelog

🆑SuhEugene
fix: Fixed a bug when robotic emote would say that emote is unusable
even if it has been successfully used.
/🆑

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->