NGR Rework Part 1: Clothing and Roles! (#2945)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

HELLO HELLO EVERYONE, IT'S THAT TIME OF THE YEAR AGAIN! ANOTHER FACTION
REWORK LET'S GOOOO WOOOOO YEAH WOOOOO!


ON THE MENU TODAY: THE NEW GORLEX REPUBLIC!


![image](https://github.com/shiptest-ss13/Shiptest/assets/77556824/b7790168-483a-4d69-9566-d13880d423de)

This PR updates ALL the visuals of our favorite boys in beige. Sadly
this does not include new ships, nor does this include new guns. You're
gonna have to wait on that for now. This PR adds three new roles that
are currently not in an NGR ship (the Hyena), which are:

- The Operative (think of it like the Enforcer)
- The Medical Doctor
- The Lieutenant (Second-In-Command role)

The Junior Agent role has now been renamed to Initiate.

Any new equipment analogous to old equipment in the Hyena has been fully
replaced, such as the old generic syndicate armor in favor of the new
suits they now boast!
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The more reworked factions, the merrier!!!!
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: PositiveEntropy
add The New Gorlex Republic's Second Battlegroup Fleet Now Boasts New
Uniforms!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->