Nerfs the size of the Etherbor Pistol and Etherbor Rifle (#2960)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

ohhh the lizard

## Why It's Good For The Game

I think they're too strong for their size, with the pistol being more or
less equal to a Sharplite laser gun that is normal sized or a pistol
that is normal sized, and the Etherbor Rifle being more or less equal to
an SMG which is also bulky sized. The BG-16 stays as it was before.

## Changelog

:cl:

balance: The civilian etherbor weapons have been made larger (the SG-8
being normal sized and the BG-12 being bulky sized)

/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->